backend/requirements/base.txt
Django>=5.0
python-dotenv>=1.0.0
Pillow>=10.0.0
django-cors-headers>=4.0.0
